+++ packages/moonstone/Scroller/Scrollable.js
@@ -107,11 +107,13 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 
 		static propTypes = /** @lends moonstone/Scroller.Scrollable.prototype */ {
 			/**
-			 * The callback function which is called for linking scrollTo function.
+			 * The callback function to get scroll top, left values.
+			 * The callback function should be called after being rendered.
+			 * If not, it will return `'{left: null, top: null}'`.
 			 * You should specify a callback function as the value of this prop
-			 * to use scrollTo feature.
+			 * to use getScrollDistance feature.
 			 *
-			 * The scrollTo function passed to the parent component requires below as an argument.
+			 * The getScrollDistance function passed to the parent component requires below as an argument.
 			 * - {position: {x, y}} - You can set a pixel value for x and/or y position
 			 * - {align} - You can set one of values below for align
 			 *   `'left'`, `'right'`, `'top'`, `'bottom'`,
@@ -119,22 +121,20 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			 * - {index} - You can set an index of specific item. (`0` or positive integer)
 			 *   This option is available for only VirtualList kind.
 			 * - {node} - You can set a node to scroll
-			 * - {animate} - When `true`, scroll occurs with animation.
-			 *   Set it to `false`, if you want scrolling without animation.
-			 * - {focus} - Set it `true`, if you want the item to be focused after scroll.
-			 *   This option is only valid when you scroll by `index` or `node`.
 			 *
 			 * Example:
 			 * ```
-			 *	// If you set cbScrollTo prop like below;
-			 *	cbScrollTo: (fn) => {this.scrollTo = fn;}
+			 *	// If you set cbScrollDistance prop like below;
+			 *	cbScrollDistance: (fn) => {this.getScrollDistance = fn;}
 			 *	// You can simply call like below;
-			 *	this.scrollTo({align: 'top'}); // scroll to the top
+			 *	const {left, top} = this.getScrollDistance({align: 'top'});
+			 *  // or
+			 *  const {left, top} = this.getScrollDistance({index: 100});
 			 * ```
 			 * @type {Function}
 			 * @public
 			 */
-			cbScrollTo: PropTypes.func,
+			cbScrollDistance: PropTypes.func,
 
 			/**
 			 * When `true`, allows 5-way navigation to the scrollbar controls. By default, 5-way will
@@ -208,7 +208,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		}
 
 		static defaultProps = {
-			cbScrollTo: nop,
+			cbScrollDistance: nop,
 			onScroll: nop,
 			onScrollStart: nop,
 			onScrollStop: nop,
@@ -252,7 +252,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				onNextScroll: this.onScrollbarButtonClick
 			};
 
-			props.cbScrollTo(this.scrollTo);
+			props.cbScrollDistance(this.getScrollDistance);
 		}
 
 		// component life cycle
@@ -1005,6 +1005,23 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 			return {left, top};
 		}
 
+		getScrollDistance = (opt) => {
+			if (this.childRef && !this.deferScrollTo) {
+				const distance = this.getPositionForScrollTo(opt);
+
+				if (distance.left === null) {
+					distance.left = this.scrollLeft;
+				}
+				if (distance.top === null) {
+					distance.top = this.scrollTop;
+				}
+
+				return distance;
+			} else {
+				return {left: null, top: null};
+			}
+		}
+
 		scrollTo = (opt) => {
 			if (!this.deferScrollTo) {
 				const {left, top} = this.getPositionForScrollTo(opt);
@@ -1220,7 +1237,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 				{isHorizontalScrollbarVisible, isVerticalScrollbarVisible} = this.state,
 				scrollableClasses = classNames(css.scrollable, className);
 
-			delete props.cbScrollTo;
+			delete props.cbScrollDistance;
 			delete props.className;
 			delete props.focusableScrollbar;
 			delete props.horizontalScrollbar;