+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unreleased]
+
+### Removed
+
+- `moonstone/VirtualList`, `moonstone/VirtualGridList`, and `moonstone/Scroller` prop `cbScrollTo` to be replaced by `cbScrollDistance`
+
 ## [2.0.0-alpha.4] - 2018-02-13
 
 ### Added